NY02 Added APIs for Ticket and Employee Services

+++ FacilitEase/Models/EntityModels/TBL_LOGIN.cs
@@ -5,7 +5,7 @@ namespace FacilitEase.Models.EntityModels
     public class TBL_LOGIN
     {
        
-        public BigInteger Id {  get; set; }
+        public long Id {  get; set; }
         public int UserId { get; set; }
         public DateTime LoginTime { get; set; }
         public DateTime LogoutTime { get; set; }
